Code review
Overall, I think we should be stricter on naming, to make the relation
between model and implementation the clearer possible.

I have two doubts about model corresponding to the documentation of the
implementation (but the tests are green):

- compare
- difference

Also, it could be nice if the failure message could display the input
so that the user can reproduce the failure.